requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2
bs4==0.0.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
fastapi==0.104.0
greenlet==3.0.0
h11==0.14.0
idna==3.4
packaging==23.2
playwright==1.39.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyee==11.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
